changing data types to float for lat and long in database

+++ migrations/20160416140645_schoolTable.js
@@ -10,8 +10,8 @@ return knex.schema.createTableIfNotExists('schoolsTable', function (table) {
   table.string("street")
   table.string("suburb")
   table.string("city")
-  table.integer("longitude")
-  table.integer("latitude")
+  table.float("longitude")
+  table.float("latitude")
 })
 
 };
